Working on Tweeting

+++ app/Console/Commands/Tweet.php
@@ -44,21 +44,23 @@ class Tweet extends Command
      */
     public function handle()
     {
-        $checks = ['server_creation_time' => "server creation time", 'speed_test_upload' => "upload speedtest", 'speed_test_download' => "download speedtest"];
-        $media = [];
-        foreach ($checks as $check => $text) {
-            $path = storage_path(str_random() . '.png');
-            Browsershot::url(env('APP_URL') . '/?hide=' . $check)->waitUntilNetworkIdle()->setDelay(1000 * 10)->fullPage()->save($path);
+        if (env('APP_LOCATION', false) != false) {
+            $checks = ['server_creation_time' => "server creation time", 'speed_test_upload' => "upload speedtest", 'speed_test_download' => "download speedtest"];
+            $media = [];
+            foreach ($checks as $check => $text) {
+                $path = storage_path(str_random() . '.png');
+                Browsershot::url(env('APP_URL') . '/?hide=' . $check)->waitUntilNetworkIdle()->setDelay(1000 * 10)->fullPage()->save($path);
+
+                $uploaded_media = Twitter::uploadMedia(['media' => File::get($path)]);
+
+                $media[] = $uploaded_media->media_id_string;
+                unlink($path);
+            }
             $providers = Provider::all()->map(function ($p) {
-                return '#' . str_repeat('-','',$p->name);
+                return '#' . str_replace('-', '', $p->name);
             });
-            $uploaded_media = Twitter::uploadMedia(['media' => File::get($path)]);
+            Twitter::postTweet(['status' => "Hi there! I've got a new daily result of my #monitoring from #" . env('APP_LOCATION') . " for you! " . $providers->implode(' ') . ' #cloud', 'media_ids' => $media]);
 
-            $media[] = $uploaded_media->media_id_string;
-            unlink($path);
         }
-
-        Twitter::postTweet(['status' => "Hi there! I've got a new daily result of my #monitoring from #" . env('APP_LOCATION') . " for you! ".$providers->implode(' ').' #cloud', 'media_ids' => $media]);
-
     }
 }
